requirements.txt
ultralytics==8.3.78
streamlit==1.42.2
moviepy==1.0.3
mlflow==2.20.3
# dvc==3.59.1
python-box==7.3.2
dagshub==0.5.8
# ipywidgets==8.1.5
setuptools==75.8.2
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
# dvc-s3==3.2.0
# uv==0.6.6
.